Created scaffold for spicetify installer
I've found out that the spicetify installer needs spotify to generate a
`~/.config/spotify/prefs` file - which it does on the first run. As the
Travis VM presumably doesn't have graphical output, the best alternative
to running spotify is to just create an empty `prefs` file - which seems
to work just fine.