Implement etckeeper
My `/etc` is now managed via the great etckeeper tools. I've got a
separate 'dotfiles-etc' repository handling all those files, that is now
just linked as a submodule into my dotfiles.

The file structure has not changed, but many extra files are now
tracked, and more will be as my system evolves.

I've still got to recover the automatic setup of /etc on install, but
that should be easy, as the `system/etc` in this repository will have
the exact same structure as `/etc`, so a simple recursive rsync should
be possible. I will do a buch of testing in a VM first though.