Major refactor - dotdrop, wayland
This is the culmination of dotfiles, that I've ended up with after 2
months of tinkering with Hyprland, and Wayland in general. Most of the
configuration files have written from scratch as I've wanted to
completely change the theme and behaviour of my environment.
I've also switched to a (in my opinion) superior dotfile manager,
Dotdrop, as it allows me to interact with the dofiles from anywhere,
manage my system level (root) dotfiles in the same place, and has a lot
of features insipred by git. It also doesn't rely on symlinks, which was
troublesome in some edge-cases before.

Overall I'm satisfied with the current state, but would like to improve
a few aspects - especially the installation procedure, package
management, and actually take advantage of the templating, and other
advanced features of dotdrop.

Signed-off-by: niraami <istok.public@gmail.com>